Generate frequency of categorical variables using table generator blocks

Generate frequency of categorical variables using table generator blocks

if ADSL supplied look for the column to take frequency of and look for a grouping variable to group_by if data is grouped add total column to the grouped data

app_y(column, group, data, totals)

# S3 method for default
app_y(column, group = NULL, data, totals)

# S3 method for BDS
app_y(column, group = NULL, data, totals)

# S3 method for custom
app_y(column, group, data, totals)

Arguments

column
the variable to perform frequency stats on, this also contains the class of the column based on the data file the column came from

group
the groups to compare for the ANOVA

data
the data to use

totals
the totals data frame that contains denominator N's use when calculating column percentages

Value

a frequency table of grouped variables

frequency table of ADSL column

See also